dockerfile/batch-processing/requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlalchemy==2.0.43
pandas==2.3.2
pyarrow==20.0.0
pyspark==3.4.2
delta-spark==2.4.0
pyyaml==5.4.1
requests==2.32.5
#uv venv .venv --python=3.8
#uv pip install -r ./dockerfile/batch-processing/requirements.txt